2.3.0 Update! Compatibility update!
Added Korean translation - thanks to @gyular
Added Ukrainian and Russian translations - thanks to @sall0-0p (_lordBucket)
Now modpack will be auto updated in preload stage of loading game so updates are faster!
Added auto_exclude_server_side_mods to config, so server-side mods will no longer be placed in the modpack which in some cases was causing a crash client
Fixed networking closing #38
Added support to velocity
Fixed zip issue
Added support for drop box, mediafire links to modpack
Quilted / Fabric Api is no longer deleted at modpack delete
TrashMod removed
Improved internet connection checks

Enjoy! :)

+++ src/main/java/pl/skidam/automodpack/AutoModpackServer.java
@@ -1,13 +1,14 @@
 package pl.skidam.automodpack;
 
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
+import com.google.gson.*;
 import net.fabricmc.api.DedicatedServerModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.networking.v1.*;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerLoginNetworkHandler;
+import net.minecraft.server.network.ServerPlayNetworkHandler;
+import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
 import org.apache.commons.io.FileUtils;
 import pl.skidam.automodpack.config.Config;
@@ -38,6 +39,8 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
     public static final File serverModsDir = new File("./mods/");
     public static String publicServerIP;
     public static File tempDir = new File("./AutoModpack/temp/");
+    public static String[] oldMods;
+    public static String[] newMods;
 
     @Override
     public void onInitializeServer() {
@@ -50,8 +53,16 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
         genModpack();
 
         // Packets
-        ServerLoginNetworking.registerGlobalReceiver(AM_LINK, this::onSuccess);
-        ServerLoginConnectionEvents.QUERY_START.register(this::onLoginStart);
+        if (!isVelocity) {
+            ServerLoginConnectionEvents.QUERY_START.register(AutoModpackServer::onLoginStart);
+            ServerLoginNetworking.registerGlobalReceiver(AM_LINK, AutoModpackServer::onSuccessLogin);
+        }
+
+        // For velocity support
+        if (isVelocity) {
+            ServerPlayConnectionEvents.JOIN.register(AutoModpackServer::onJoinStart);
+            ServerPlayNetworking.registerGlobalReceiver(AM_LINK, AutoModpackServer::onSuccessJoin);
+        }
 
         if (modpackZip.exists()) {
             ServerLifecycleEvents.SERVER_STARTED.register(server -> HostModpack.start());
@@ -59,25 +70,87 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
         }
     }
 
+    private static void onSuccessJoin(MinecraftServer minecraftServer, ServerPlayerEntity serverPlayerEntity, ServerPlayNetworkHandler serverPlayNetworkHandler, PacketByteBuf buf, PacketSender sender) {
+        // Successfully sent link to client, client can join and play on server.
+        if (!buf.readString().equals("1")) {
+            if (!Config.ONLY_OPTIONAL_MODPACK) { // Accept player to join while optional modpack is enabled
+                serverPlayNetworkHandler.disconnect(Text.of("You have to install \"AutoModpack\" mod to play on this server! https://modrinth.com/mod/automodpack/versions"));
+                LOGGER.warn("Player " + serverPlayNetworkHandler.getPlayer().getName().getString() + " has not installed \"AutoModpack\" mod");
+            } else {
+                LOGGER.warn("Player " + serverPlayNetworkHandler.getPlayer().getName().getString()  + " has not installed modpack");
+            }
+        }
+    }
+
+    private static void onJoinStart(ServerPlayNetworkHandler serverPlayNetworkHandler, PacketSender sender, MinecraftServer minecraftServer) {
+        // Get minecraft player ip if player is in local network give him local address to modpack
+        String playerIp = serverPlayNetworkHandler.getConnection().getAddress().toString();
+        PacketByteBuf outBuf = PacketByteBufs.create();
+        if (playerIp.contains("127.0.0.1") || playerIp.contains(publicServerIP)) {
+            outBuf.writeString(HostModpack.modpackHostIpForLocalPlayers);
+        } else {
+            outBuf.writeString(AutoModpackMain.link);
+        }
+        sender.sendPacket(AutoModpackMain.AM_LINK, outBuf);
+        LOGGER.info("Sent link to " + serverPlayNetworkHandler.getPlayer().getName().getString() + " through velocity");
+    }
+
+    private static void onLoginStart(ServerLoginNetworkHandler serverLoginNetworkHandler, MinecraftServer minecraftServer, PacketSender sender, ServerLoginNetworking.LoginSynchronizer loginSynchronizer) {
+        // Get minecraft player ip if player is in local network give him local address to modpack
+        String playerIp = serverLoginNetworkHandler.getConnection().getAddress().toString();
+        PacketByteBuf outBuf = PacketByteBufs.create();
+        if (playerIp.contains("127.0.0.1") || playerIp.contains(publicServerIP)) {
+            outBuf.writeString(HostModpack.modpackHostIpForLocalPlayers);
+        } else {
+            outBuf.writeString(AutoModpackMain.link);
+        }
+        sender.sendPacket(AutoModpackMain.AM_LINK, outBuf);
+        LOGGER.info("Sent link to " + getPlayerNickInLogin(serverLoginNetworkHandler.getConnectionInfo()));
+    }
+
+    private static void onSuccessLogin(MinecraftServer minecraftServer, ServerLoginNetworkHandler serverLoginNetworkHandler, boolean understood, PacketByteBuf buf, ServerLoginNetworking.LoginSynchronizer loginSynchronizer, PacketSender sender) {
+        // Successfully sent link to client, client can join and play on server.
+        if (!understood || !buf.readString().equals("1")) {
+            if (!Config.ONLY_OPTIONAL_MODPACK) { // Accept player to join while optional modpack is enabled
+                serverLoginNetworkHandler.disconnect(Text.of("You have to install \"AutoModpack\" mod to play on this server! https://modrinth.com/mod/automodpack/versions"));
+                LOGGER.warn("Player " + getPlayerNickInLogin(serverLoginNetworkHandler.getConnectionInfo()) + " has not installed \"AutoModpack\" mod");
+            } else {
+                LOGGER.warn("Player " + getPlayerNickInLogin(serverLoginNetworkHandler.getConnectionInfo()) + " has not installed modpack");
+            }
+        }
+    }
+
+    private static String getPlayerNickInLogin(String connectionInfo) {
+        String nick;
+        try {
+            nick = connectionInfo.split("name=")[1];
+            nick = nick.split(",")[0];
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+
+        return nick;
+    }
+
     public static void genModpack() {
 
-        clientMods();
-
         if (!Config.SYNC_MODS) {
-            onlyServerSideMods();
+            autoExcludeMods();
         }
+
+        clientMods();
 
         // Sync mods and automatically generate delmods.txt
         if (Config.SYNC_MODS) {
             LOGGER.info("Synchronizing mods from server to modpack");
 
             // Make array of mods
-            String[] oldMods = modpackModsDir.list();
+            oldMods = modpackModsDir.list();
             deleteAllMods();
             cloneMods();
+            autoExcludeMods();
             clientMods();
-            onlyServerSideMods();
-            String[] newMods = modpackModsDir.list();
+            newMods = modpackModsDir.list();
 
             // Changelog generation
             SimpleDateFormat ft = new SimpleDateFormat("yyyy-MM-dd");
@@ -251,9 +324,9 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
         }
     }
 
-    private static void onlyServerSideMods() {
+    private static void autoExcludeMods() {
         if (!Config.AUTO_EXCLUDE_SERVER_SIDE_MODS) return;
-        LOGGER.info("Excluding server side mods...");
+        LOGGER.info("Excluding server-side mods from modpack");
         for (File file : Objects.requireNonNull(serverModsDir.listFiles())) {
             if (file.getName().endsWith(".jar") && !file.getName().toLowerCase().contains("automodpack")) {
                 try {
@@ -271,16 +344,17 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
                 if (serverSideMods == null) return;
 
                 if (new File(tempDir + "/" + file.getName()).exists()) {
-                    File serverSideModJson = null;
+                    File modJsonFile = null;
                     if (new File(tempDir + "/" + file.getName() + "/fabric.mod.json").exists()) {
-                        serverSideModJson = new File(tempDir + "/" + file.getName() + "/fabric.mod.json");
+                        modJsonFile = new File(tempDir + "/" + file.getName() + "/fabric.mod.json");
                     } else if (new File(tempDir + "/" + file.getName() + "/quilt.mod.json").exists()) {
-                        serverSideModJson = new File(tempDir + "/" + file.getName() + "/quilt.mod.json");
+                        modJsonFile = new File(tempDir + "/" + file.getName() + "/quilt.mod.json");
                     }
 
-                    if (serverSideModJson != null) {
+                    if (modJsonFile != null) {
                         try {
-                            JsonObject jsonObject = JsonParser.parseReader(new FileReader(serverSideModJson)).getAsJsonObject();
+                            FileReader modJsonReader = new FileReader(modJsonFile);
+                            JsonObject jsonObject = JsonParser.parseReader(modJsonReader).getAsJsonObject();
                             String environment;
                             try {
                                 environment = jsonObject.get("environment").getAsString();
@@ -288,18 +362,40 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
                                 try {
                                     JsonObject quilt_loader = jsonObject.get("quilt_loader").getAsJsonObject(); // for quilt support
                                     environment = quilt_loader.get("environment").getAsString();
-                                } catch (Exception e2) { // this mod doesn't have provided any environment
-                                    continue;
+                                } catch (Exception e2) { // this mod doesn't have provided any environment lol
+                                    environment = "*";
                                 }
                             }
-                            if (environment.equals("server")) {
+                            modJsonReader.close();
+                            if (Config.AUTO_EXCLUDE_SERVER_SIDE_MODS && environment.equals("server")) {
                                 File serverSideModInModpack = new File(modpackModsDir + "/" + file.getName());
                                 if (serverSideModInModpack.exists()) {
                                     FileUtils.deleteQuietly(serverSideModInModpack);
-                                    LOGGER.info(file.getName() + " is server side mod and has been auto excluded from modpack");
+                                    for (String oldMod : oldMods) { // log to console if this mod was in modpack before
+                                        if (oldMod.equals(serverSideModInModpack.getName())) {
+                                            LOGGER.info(file.getName() + " is server-side mod and has been auto excluded from modpack");
+                                        }
+                                    }
                                 }
                             }
-                        } catch (FileNotFoundException e) {
+
+                            // BRAIN LAG THIS IS NON SENSE
+
+//                            if (Config.AUTO_EXCLUDE_CLIENT_SIDE_MODS && environment.equals("client")) {
+//                                // copy file to client side mods in modpack
+//                                File clientSideModInModpack = new File(modpackClientModsDir + "/" + file.getName());
+//                                if (!clientSideModInModpack.exists() || clientSideModInModpack.length() != file.length()) {
+//                                    FileUtils.deleteQuietly(clientSideModInModpack);
+//                                    FileUtils.copyFileToDirectory(file, modpackClientModsDir);
+//                                }
+//                                for (String oldMod : oldMods) { // log to console if this mod was in modpack before
+//                                    if (!oldMod.equals(clientSideModInModpack.getName())) {
+//                                        LOGGER.info(file.getName() + " is client side mod and has been auto excluded from server mods");
+//                                    }
+//                                }
+//                                FileUtils.deleteQuietly(file);
+//                            }
+                        } catch (IOException e) {
                             e.printStackTrace();
                         }
                     }
@@ -307,6 +403,7 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
                 }
             }
         }
+        FileUtils.deleteQuietly(tempDir);
     }
 
     private static void clientMods() {
@@ -328,36 +425,5 @@ public class AutoModpackServer implements DedicatedServerModInitializer {
                 LOGGER.error("Error while deleting the file: " + file);
             }
         }
-    }
-
-    private void onSuccess(MinecraftServer minecraftServer, ServerLoginNetworkHandler serverLoginNetworkHandler, boolean understood, PacketByteBuf buf, ServerLoginNetworking.LoginSynchronizer loginSynchronizer, PacketSender sender) {
-        // Successfully sent link to client, client can join and play on server.
-        if (!understood || !buf.readString().equals("1")) {
-            if (!Config.ONLY_OPTIONAL_MODPACK) { // Accept player to join while optional modpack is enabled // TODO make it better
-                serverLoginNetworkHandler.disconnect(Text.of("You have to install \"AutoModpack\" mod to play on this server! https://modrinth.com/mod/automodpack/versions"));
-                LOGGER.warn("Player " + serverLoginNetworkHandler.getConnectionInfo() + " has not installed \"AutoModpack\" mod");
-            } else {
-                LOGGER.warn("Player " + serverLoginNetworkHandler.getConnectionInfo() + " has not installed modpack");
-                serverLoginNetworkHandler.acceptPlayer();
-            }
-        } else {
-            LOGGER.warn("Player " + serverLoginNetworkHandler.getConnectionInfo() + " has installed modpack");
-            serverLoginNetworkHandler.acceptPlayer();
-        }
-    }
-
-    private void onLoginStart(ServerLoginNetworkHandler serverLoginNetworkHandler, MinecraftServer minecraftServer, PacketSender sender, ServerLoginNetworking.LoginSynchronizer loginSynchronizer) {
-        // Get minecraft player ip if player is in local network give him local address to modpack
-        String playerIp = serverLoginNetworkHandler.getConnection().getAddress().toString();
-
-        PacketByteBuf outBuf = PacketByteBufs.create();
-
-        if (playerIp.contains("127.0.0.1") || playerIp.contains(publicServerIP)) {
-            outBuf.writeString(HostModpack.modpackHostIpForLocalPlayers);
-        } else {
-            outBuf.writeString(AutoModpackMain.link);
-        }
-
-        sender.sendPacket(AutoModpackMain.AM_LINK, outBuf);
     }
 }
